Revert "Add batch badge requests"
