+++ AntGameAPI/app.js
@@ -42,6 +42,7 @@ const UnauthenticatedRoutes = [
   /\/flag\//,
   /\/service\//,
   /\/public\//,
+  /\/user\/[A-z0-9]*\/badges/,
   "/health",
   "/time",
 ];
@@ -132,13 +133,12 @@ app.get("/public/activeChallenges", _publicController.getActiveChallenges);
 app.get("/public/challengeLeaderboard/:id", _publicController.getChallengeLeaderboard);
 app.get("/public/dailyList", _publicController.getDailyChallenges);
 app.get("/public/gsgp", _publicController.getGsgpData);
-app.get("/public/badges/:id", _publicController.getUserBadges);
 
 app.post("/seed", getSeedLimiter, _seedController.getSeed);
 
 app.get("/championship/:id", _championshipController.getLeaderboard);
 
-app.post("/badges", _userController.getUserBadges);
+app.get("/user/:id/badges", _userController.getUserBadges);
 
 app.post("/report/spaces", _reportController.reportSpacesData);
