+++ AntGameUI/src/antGame/User/UserService.js
@@ -1,9 +1,5 @@
 import axios from "axios";
 
 export const GetUserBadges = async id => {
-  return await axios.get(`/api/public/badges/${id}`).then(res => res.data);
-};
-
-export const GetBatchBadges = async userList => {
-  return await axios.post(`/api/badges`, { userList }).then(res => res.data);
+  return await axios.get(`/api/user/${id}/badges`).then(res => res.data);
 };
